+++ src/executor/engine.rs
@@ -202,10 +202,10 @@ impl<'a> ExecutionContext<'a> {
         let right_rows = self.resolve_table_source(&join.table)?;
 
         match join.join_type {
-            JoinType::Inner | JoinType::Straight => {
+            JoinType::Inner | JoinType::Straight | JoinType::CrossApply => {
                 self.inner_join(left_rows, &right_rows, &join.on, &join.using)
             }
-            JoinType::Left | JoinType::LeftOuter => {
+            JoinType::Left | JoinType::LeftOuter | JoinType::OuterApply => {
                 self.left_join(left_rows, &right_rows, &join.on, &join.using)
             }
             JoinType::Right | JoinType::RightOuter => {